requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
langchain_community==0.3.5
pdf2image==1.17.0
Pillow==11.0.0
Werkzeug==3.1.3
WTForms==3.2.1
